modifying to change floating point to interger

+++ rfmix_reader/_fb_reader.h
@@ -4,6 +4,6 @@
  * This is modified to handle a matrix of floating-point numbers.
  */
 
-void read_fb_chunk(float *buff, uint64_t nrows, uint64_t ncols,
+void read_fb_chunk(uint8_t *buff, uint64_t nrows, uint64_t ncols,
 		   uint64_t row_start, uint64_t col_start, uint64_t row_end,
-		   uint64_t col_end, float *out, uint64_t *strides);
+		   uint64_t col_end, uint8_t *out, uint64_t *strides);
